requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.0
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-richtextfield==1.6.1
gunicorn==20.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
